Frontend/requirements.txt
from flask_restful import Resource
from flask import request
python-dotenv
Flask